Added abstraction layer to pthread function calls.
Since the goal is to have the ivi-logging also compiling on an Android
environment, an abstraction layer is provided for pthread calls. This
allows to consider if either or not a given pthread functionality is
available on the current android platform. For instance
pthread_getname_np is only available since __ANDROID_API__ >= 26